# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: vcdeploy API
  description: |
    REST API for vcdeploy - a modern deployment orchestration platform.
    
    ## Authentication
    
    Most endpoints require authentication via one of:
    - Session cookie (for web UI)
    - Bearer token in `Authorization` header (for API access)
    - API key in `X-API-Key` header
    
    ## Rate Limiting
    
    API requests are rate limited. When exceeded, endpoints return `429 Too Many Requests`.
    
    ## Request IDs
    
    All requests receive an `X-Request-ID` header for correlation and debugging.
  version: 1.0.0
  license:
    name: MIT
    url: https://github.com/BlackOrder/vcdeploy/blob/main/LICENSE
  contact:
    name: BlackOrder
    url: https://github.com/BlackOrder/vcdeploy
servers:
  - url: http://localhost:9000
    description: Local development server
  - url: https://vcdeploy.example.com
    description: Production server (example)
...
paths:
...
  /api/v1/host-keys:
    get:
      tags: [Agents]
      summary: List SSH host keys
      description: Returns all known SSH host keys.
      operationId: listHostKeys
      responses:
        '200':
          description: List of host keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HostKey'
    post:
      tags: [Agents]
      summary: Add SSH host key
      description: Add a new SSH host key to the known hosts.
      operationId: createHostKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostKeyCreate'
      responses:
        '201':
          description: Host key added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostKey'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/jump-servers:
    get:
      tags: [Agents]
      summary: List jump servers
      description: Returns all configured jump servers (bastion hosts).
      operationId: listJumpServers
      responses:
        '200':
          description: List of jump servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JumpServer'
    post:
      tags: [Agents]
      summary: Add jump server
      description: Add a new jump server configuration.
      operationId: createJumpServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JumpServerCreate'
      responses:
        '201':
          description: Jump server added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JumpServer'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/jump-servers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: Jump server ID
    get:
      tags: [Agents]
      summary: Get jump server
      description: Get details of a specific jump server.
      operationId: getJumpServer
      responses:
        '200':
          description: Jump server details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JumpServer'
        '404':
          description: Jump server not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Agents]
      summary: Update jump server
      description: Update a jump server configuration.
      operationId: updateJumpServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JumpServerCreate'
      responses:
        '200':
          description: Jump server updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JumpServer'
        '404':
          description: Jump server not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Agents]
      summary: Delete jump server
      description: Remove a jump server configuration.
      operationId: deleteJumpServer
      responses:
        '204':
          description: Jump server deleted
        '404':
          description: Jump server not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/blocked:
    get:
      tags: [Settings]
      summary: List blocked IPs
      description: Returns all currently blocked IP addresses.
      operationId: listBlockedIPs
      responses:
        '200':
          description: List of blocked IPs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BlockedIP'
    post:
      tags: [Settings]
      summary: Block IP address
      description: Manually block an IP address.
      operationId: blockIP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockIPRequest'
      responses:
        '201':
          description: IP blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockedIP'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/blocked/{ip}:
    parameters:
      - name: ip
        in: path
        required: true
        schema:
          type: string
        description: IP address to manage
    get:
      tags: [Settings]
      summary: Get blocked IP details
      description: Get details of a specific blocked IP.
      operationId: getBlockedIP
      responses:
        '200':
          description: Blocked IP details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockedIP'
        '404':
          description: IP not found in block list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Settings]
      summary: Unblock IP address
      description: Remove an IP address from the block list.
      operationId: unblockIP
      responses:
        '204':
          description: IP unblocked
        '404':
          description: IP not found in block list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/provision:
    get:
      tags: [Agents]
      summary: List provision jobs
      description: Returns all server provisioning jobs.
      operationId: listProvisionJobs
      responses:
        '200':
          description: List of provision jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProvisionJob'
    post:
      tags: [Agents]
      summary: Create provision job
      description: Create a new server provisioning job.
      operationId: createProvisionJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisionJobCreate'
      responses:
        '201':
          description: Provision job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvisionJob'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key

  # H8 FIX: Reusable pagination parameters
...
  schemas:
...
    Error:
      type: object
      description: Standard error response for all API endpoints
      required:
        - error
        - message
      properties:
        error:
          type: boolean
          description: Always true for error responses
          example: true
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"

    # H8 FIX: Add pagination metadata schema
...
    HostKey:
      type: object
      properties:
        id:
          type: integer
        hostname:
          type: string
        port:
          type: integer
        keyType:
          type: string
        publicKey:
          type: string
        fingerprint:
          type: string
        trusted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastVerifiedAt:
          type: string
          format: date-time
    HostKeyCreate:
      type: object
      required: [hostname, keyType, publicKey]
      properties:
        hostname:
          type: string
        port:
          type: integer
          default: 22
        keyType:
          type: string
        publicKey:
          type: string
        fingerprint:
          type: string
        trusted:
          type: boolean
          default: false
    JumpServer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        hostname:
          type: string
        port:
          type: integer
        username:
          type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    JumpServerCreate:
      type: object
      required: [name, hostname, username]
      properties:
        name:
          type: string
        hostname:
          type: string
        port:
          type: integer
          default: 22
        username:
          type: string
        privateKey:
          type: string
          format: password
        enabled:
          type: boolean
          default: true
    BlockedIP:
      type: object
      properties:
        id:
          type: integer
        ipAddress:
          type: string
        blockedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        reason:
          type: string
        manualBlock:
          type: boolean
    BlockIPRequest:
      type: object
      required: [ipAddress]
      properties:
        ipAddress:
          type: string
        reason:
          type: string
        duration:
          type: string
          description: Duration string (e.g., "1h", "24h", "7d")
    ProvisionJob:
      type: object
      properties:
        id:
          type: string
        targetHost:
          type: string
        targetPort:
          type: integer
        targetUser:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        output:
          type: string
        errorMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    ProvisionJobCreate:
      type: object
      required: [targetHost, targetUser]
      properties:
        targetHost:
          type: string
        targetPort:
          type: integer
          default: 22
        targetUser:
          type: string
        jumpServerId:
          type: integer
          description: Optional jump server to use
...
security:
  - bearerAuth: []
  - apiKey: []